automobile_service/requirements.txt
Django>=3.2,<4.0
djangorestframework
gunicorn
psycopg2-binary
celery
django-environ==0.11.2
